a_3.cpp change int R -> double R

+++ op/lab_01/a_3.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 
 int main() {
-    int R;
+    double R;
     double pi = 3.14159;
     std::cin >> R;
